+++ issuer-application/src/main/java/ch/admin/bj/swiyu/issuer/infrastructure/web/signer/WellKnownController.java
@@ -8,18 +8,16 @@ package ch.admin.bj.swiyu.issuer.infrastructure.web.signer;
 
 import ch.admin.bj.swiyu.issuer.api.oid4vci.OpenIdConfigurationDto;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.IssuerMetadata;
-import ch.admin.bj.swiyu.issuer.infrastructure.config.OpenIdIssuerApiConfiguration;
 import ch.admin.bj.swiyu.issuer.service.EncryptionService;
+import ch.admin.bj.swiyu.issuer.service.MetadataService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.aop.framework.AopProxyUtils;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
+import java.util.UUID;
 
 /**
  * Well known Controller
@@ -35,21 +33,22 @@ import java.io.IOException;
         "and credential metadata as required by the OID4VCI specification. Provides endpoints for OpenID Connect " +
         "issuer configuration, OAuth authorization server information, and issuer metadata describing supported " +
         "verifiable credentials (IF-112)")
-@RequestMapping(value = {"/oid4vci/.well-known", ".well-known"})
+@RequestMapping
 public class WellKnownController {
 
-    private final OpenIdIssuerApiConfiguration openIDConfiguration;
+    private static String CONTENT_TYPE_APPLICATION_JWT = "application/jwt";
     private final EncryptionService encryptionService;
+    private final MetadataService metadataService;
 
     /**
      * General information about the issuer
      *
      * @return OpenIdConfigurationDto as defined by OIDConnect and extended by OID4VCI
      */
-    @GetMapping("/openid-configuration")
+    @GetMapping(value = {"/oid4vci/.well-known/openid-configuration", ".well-known/openid-configuration"})
     @Operation(summary = "OpenID Connect information required for issuing VCs")
-    public OpenIdConfigurationDto getOpenIDConfiguration() throws IOException {
-        return openIDConfiguration.getOpenIdConfiguration();
+    public OpenIdConfigurationDto getOpenIDConfiguration() {
+        return metadataService.getUnsignedOpenIdConfiguration();
     }
 
     /**
@@ -58,10 +57,10 @@ public class WellKnownController {
      *
      * @return OpenIdConfigurationDto as defined by OIDConnect and extended by OID4VCI
      */
-    @GetMapping("/oauth-authorization-server")
+    @GetMapping(value = {"/oid4vci/.well-known/oauth-authorization-server", ".well-known/oauth-authorization-server"})
     @Operation(summary = "OpenID Connect information required for issuing VCs")
-    public OpenIdConfigurationDto getOpenIDConfigurationForOauthAuthServer() throws IOException {
-        return openIDConfiguration.getOpenIdConfiguration();
+    public OpenIdConfigurationDto getOpenIDConfigurationForOauthAuthServer() {
+        return metadataService.getUnsignedOpenIdConfiguration();
     }
 
     /**
@@ -69,10 +68,36 @@ public class WellKnownController {
      *
      * @return Issuer Metadata as defined by OID4VCI
      */
-    @GetMapping(value = {"/openid-credential-issuer"})
+    @GetMapping(value = {"/oid4vci/.well-known/openid-credential-issuer", ".well-known/openid-credential-issuer"})
     @Operation(summary = "Information about credentials which can be issued.")
     public IssuerMetadata getIssuerMetadata() {
         // Unwrap the object from the spring cache object.
         return (IssuerMetadata) AopProxyUtils.getSingletonTarget(encryptionService.issuerMetadataWithEncryptionOptions());
+    }
+
+    @GetMapping(value = {"/{tenantId}/.well-known/openid-credential-issuer", "/oid4vci/{tenantId}/.well-known/openid-credential-issuer"})
+    @Operation(summary = "Information about credentials which can be issued.")
+    public Object getIssuerMetadataByTenantId(
+            @PathVariable UUID tenantId,
+            @RequestHeader("Content-Type") String contentType) {
+
+        if (CONTENT_TYPE_APPLICATION_JWT.equals(contentType)) {
+            return metadataService.getSignedIssuerMetadata(tenantId);
+        }
+
+        return metadataService.getUnsignedIssuerMetadata();
+    }
+
+    @GetMapping(value = {"/{tenantId}/.well-known/openid-configuration", "/oid4vci/{tenantId}/.well-known/openid-configuration"})
+    @Operation(summary = "Information about credentials which can be issued.")
+    public Object getOpenIdConfigurationByTenantId(
+            @PathVariable UUID tenantId,
+            @RequestHeader("Content-Type") String contentType) {
+
+        if (contentType.startsWith(CONTENT_TYPE_APPLICATION_JWT)) {
+            return metadataService.getSignedOpenIdConfiguration(tenantId);
+        }
+
+        return metadataService.getUnsignedOpenIdConfiguration();
     }
 }