+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/common/config/ApplicationProperties.java
@@ -102,6 +102,7 @@ public class ApplicationProperties {
     private String dataIntegrityJwks;
     private JWKSet dataIntegrityKeySet;
     private boolean dataIntegrityEnforced;
+    private boolean signedMetadataEnabled;
 
     @PostConstruct
     public void init() {