+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/OpenIdIssuerConfiguration.java
@@ -1,15 +1,10 @@
-/*
- * SPDX-FileCopyrightText: 2025 Swiss Confederation
- *
- * SPDX-License-Identifier: MIT
- */
-
-package ch.admin.bj.swiyu.issuer.infrastructure.config;
+package ch.admin.bj.swiyu.issuer.service;
 
 import ch.admin.bj.swiyu.issuer.api.oid4vci.OpenIdConfigurationDto;
 import ch.admin.bj.swiyu.issuer.api.type_metadata.OcaDto;
 import ch.admin.bj.swiyu.issuer.api.type_metadata.TypeMetadataDto;
 import ch.admin.bj.swiyu.issuer.common.config.ApplicationProperties;
+import ch.admin.bj.swiyu.issuer.common.exception.ConfigurationException;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.CredentialMetadata;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.IssuerMetadata;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -32,18 +27,15 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static ch.admin.bj.swiyu.issuer.common.config.CacheConfig.ISSUER_METADATA_PARSED_CACHE;
-import static ch.admin.bj.swiyu.issuer.common.config.CacheConfig.OPEN_ID_CONFIGURATION_CACHE;
+import static ch.admin.bj.swiyu.issuer.common.config.CacheConfig.*;
 
 @Configuration
 @Data
 @Slf4j
-public class OpenIdIssuerApiConfiguration {
+public class OpenIdIssuerConfiguration {
 
     private final ApplicationProperties applicationProperties;
-
     private final ResourceLoader resourceLoader;
-
     private final ObjectMapper objectMapper;
 
     @Value("${application.openid-file}")
@@ -51,16 +43,6 @@ public class OpenIdIssuerApiConfiguration {
 
     @Value("${application.metadata-file}")
     private Resource issuerMetadataResource;
-
-    @Cacheable(OPEN_ID_CONFIGURATION_CACHE)
-    public OpenIdConfigurationDto getOpenIdConfiguration() throws IOException {
-        return resourceToMappedData(openIdResource, OpenIdConfigurationDto.class);
-    }
-
-    @Cacheable(ISSUER_METADATA_PARSED_CACHE)
-    public IssuerMetadata getIssuerMetadata() throws IOException {
-        return resourceToMappedData(issuerMetadataResource, IssuerMetadata.class);
-    }
 
     /**
      * @return Issuer Metadata for using in creation of a vc
@@ -91,6 +73,45 @@ public class OpenIdIssuerApiConfiguration {
             builder.jsonSchemaMap(loadMetadataFiles(applicationProperties.getJsonSchemaMetadataFiles(), validator, null));
             builder.overlayCaptureArchitectureMap(loadMetadataFiles(applicationProperties.getOverlaysCaptureArchitectureMetadataFiles(), validator, OcaDto.class));
             return builder.build();
+        }
+    }
+
+    /**
+     * @return the full Issuer Metadata in a recursive Map
+     */
+    @Cacheable(ISSUER_METADATA_CACHE)
+    public IssuerMetadata getIssuerMetadata() {
+        try {
+            return resourceToMappedData(issuerMetadataResource, IssuerMetadata.class);
+        } catch (IOException e) {
+            throw new ConfigurationException("Cannot read issuer metadata", e);
+        }
+    }
+
+    @Cacheable(OPEN_ID_CONFIGURATION_CACHE)
+    public OpenIdConfigurationDto getOpenIdConfiguration() {
+        try {
+            return resourceToMappedData(openIdResource, OpenIdConfigurationDto.class);
+        } catch (IOException e) {
+            throw new ConfigurationException("Cannot read openid config", e);
+        }
+    }
+
+    @Cacheable(ISSUER_METADATA_MAP_CACHE)
+    public Map<String, Object> getIssuerMetadataMap() {
+        try {
+            return resourceToMappedData(issuerMetadataResource, HashMap.class);
+        } catch (IOException e) {
+            throw new ConfigurationException("Cannot read issuer metadata config", e);
+        }
+    }
+
+    @Cacheable(OPEN_ID_CONFIGURATION_MAP_CACHE)
+    public Map<String, Object> getOpenIdConfigurationMap() {
+        try {
+            return resourceToMappedData(openIdResource, HashMap.class);
+        } catch (IOException e) {
+            throw new ConfigurationException("Cannot read issuer metadata config", e);
         }
     }
 
@@ -143,15 +164,8 @@ public class OpenIdIssuerApiConfiguration {
         return new ObjectMapper().readValue(json, clazz);
     }
 
-    /**
-     * Loads metadata, replacing placeholders
-     *
-     * @param res resource from which the data is loaded
-     * @return the loaded metadata
-     */
     private String loadMetadata(Resource res) throws IOException {
         var json = res.getContentAsString(Charset.defaultCharset());
         return replaceExternalUri(json);
     }
-
 }