+++ issuer-service/src/test/java/ch/admin/bj/swiyu/issuer/oid4vci/test/TestServiceUtils.java
@@ -98,6 +98,7 @@ public class TestServiceUtils {
                 null,
                 null,
                 null,
+                null,
                 Instant.now().plusSeconds(600).getEpochSecond(),
                 nonce,
                 preAuthorizedCode,